requirements.txt
amqp==5.1.1
asgiref==3.7.2
billiard==4.1.0
celery==5.3.4
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
dj-rest-auth==5.0.1
Django==4.2.5
django-filter==23.3
djangorestframework==3.14.0
flower==2.0.1
humanize==4.8.0
kombu==5.3.2
Markdown==3.4.4
mypy-extensions==1.0.0
packaging==23.2
pathspec==0.11.2
Pillow==10.0.1
platformdirs==3.10.0
prometheus-client==0.17.1
prompt-toolkit==3.0.39
psycopg2-binary==2.9.8
python-dateutil==2.8.2
pytz==2023.3.post1
redis==5.0.1
six==1.16.0
sqlparse==0.4.4
tornado==6.3.3
tzdata==2023.3
vine==5.0.0
wcwidth==0.2.7
